// src/controllers/enums.ts
// Generated by format-carbon carbon-class --emit-enums from schema enums.json.

export interface DestinationTypeEnum {
  readonly OWNER: 0;
  readonly CHILD: 1;
}

export type DestinationTypeValue =
  DestinationTypeEnum[keyof DestinationTypeEnum];

export const DestinationType: DestinationTypeEnum = Object.freeze({
  OWNER: 0,
  CHILD: 1,
});

export interface PlayActionEnum {
  readonly PLAY: 0;
  readonly ENQUEUE_PLAY: 1;
}

export type PlayActionValue = PlayActionEnum[keyof PlayActionEnum];

export const PlayAction: PlayActionEnum = Object.freeze({
  PLAY: 0,
  ENQUEUE_PLAY: 1,
});

export interface ResetBehaviorEnum {
  readonly OBJECT_CENTER: 0;
  readonly LAST_DAMAGELOCATOR_HIT: 1;
  readonly CUSTOM: 2;
}

export type ResetBehaviorValue = ResetBehaviorEnum[keyof ResetBehaviorEnum];

export const ResetBehavior: ResetBehaviorEnum = Object.freeze({
  OBJECT_CENTER: 0,
  LAST_DAMAGELOCATOR_HIT: 1,
  CUSTOM: 2,
});

export interface StopActionEnum {
  readonly STOP: 0;
  readonly ENQUEUE_STOP: 1;
  readonly NONE: 2;
}

export type StopActionValue = StopActionEnum[keyof StopActionEnum];

export const StopAction: StopActionEnum = Object.freeze({
  STOP: 0,
  ENQUEUE_STOP: 1,
  NONE: 2,
});

export interface TypeEnum {
  readonly FLOAT: 0;
  readonly INTEGER: 1;
  readonly BOOLEAN: 2;
  readonly ENUM: 3;
}

export type TypeValue = TypeEnum[keyof TypeEnum];

export const Type: TypeEnum = Object.freeze({
  FLOAT: 0,
  INTEGER: 1,
  BOOLEAN: 2,
  ENUM: 3,
});

export interface UnlinkReasonEnum {
  readonly UNLINKING: 0;
  readonly DELETING: 1;
}

export type UnlinkReasonValue = UnlinkReasonEnum[keyof UnlinkReasonEnum];

export const UnlinkReason: UnlinkReasonEnum = Object.freeze({
  UNLINKING: 0,
  DELETING: 1,
});